+++ e2e/next-sandbox/test/utils.ts
@@ -45,15 +45,24 @@ export function genRoomId(testInfo: TestInfo) {
   return roomId;
 }
 
-export async function preparePage(url: string, windowPositionX: number = 0) {
+type WindowOptions = {
+  x?: number;
+  y?: number;
+  width?: number;
+  height?: number;
+};
+
+export async function preparePage(url: string, options?: WindowOptions) {
+  const width = options?.width ?? WIDTH;
+  const height = options?.height ?? HEIGHT;
   const browser = await chromium.launch({
     args: [
-      `--window-size=${WIDTH},${HEIGHT}`,
-      `--window-position=${windowPositionX},0`,
+      `--window-size=${width},${height}`,
+      `--window-position=${options?.x ?? 0},${options?.y ?? 0}`,
     ],
   });
   const context = await browser.newContext({
-    viewport: { width: WIDTH, height: HEIGHT },
+    viewport: { width, height },
   });
   const page = await context.newPage();
   await page.goto(url);
@@ -92,12 +101,21 @@ const BG_COLORS = [
   "#faf9ff",
 ];
 
+type PreparePagesOptions = WindowOptions & {
+  /** How many windows to open */
+  n?: number;
+};
+
 export async function preparePages(url: string): Promise<[Page, Page]>;
-export async function preparePages(url: string, n: number): Promise<Page[]>;
 export async function preparePages(
   url: string,
-  n: number = 2
+  options?: PreparePagesOptions
+): Promise<Page[]>;
+export async function preparePages(
+  url: string,
+  options?: PreparePagesOptions
 ): Promise<Page[]> {
+  const n = options?.n ?? 2;
   return Promise.all(
     Array.from({ length: n }).map((_, index) => {
       const pageUrl = new URL(url);
@@ -105,7 +123,10 @@ export async function preparePages(
 
       // If n=2, open the windows side-by-side, otherwise open them as a fan
       const xPos = n <= 2 && index === 1 ? WIDTH : index * 50;
-      return preparePage(pageUrl.toString(), xPos);
+      return preparePage(pageUrl.toString(), {
+        ...options,
+        x: options?.x ?? xPos,
+      });
     })
   );
 }
@@ -113,16 +134,22 @@ export async function preparePages(
 export async function waitForJson(
   oneOrMorePages: Page | Page[],
   selector: IDSelector,
-  expectedValue: Json
+  expectedValue: Json | undefined,
+  options?: { timeout?: number }
 ) {
   const pages = Array.isArray(oneOrMorePages)
     ? oneOrMorePages
     : [oneOrMorePages];
 
-  const expectedText = JSON.stringify(expectedValue, null, 2);
+  const expectedText =
+    expectedValue === undefined
+      ? "undefined"
+      : JSON.stringify(expectedValue, null, 2);
   return Promise.all(
     pages.map((page) =>
-      expect(page.locator(selector)).toHaveText(expectedText, { timeout: 5000 })
+      expect(page.locator(selector)).toHaveText(expectedText, {
+        timeout: options?.timeout ?? 5000,
+      })
     )
   );
 }
@@ -140,12 +167,15 @@ export async function expectJson(
   }
 }
 
-export async function getJson(page: Page, selector: IDSelector): Promise<Json> {
+export async function getJson(
+  page: Page,
+  selector: IDSelector
+): Promise<Json | undefined> {
   const text = await page.locator(selector).innerText();
   if (!text) {
     throw new Error(`Could not find HTML element #${selector}`);
   }
-  return JSON.parse(text) as Json;
+  return text === "undefined" ? undefined : (JSON.parse(text) as Json);
 }
 
 async function getBoth(pages: [Page, Page], selector: IDSelector) {
@@ -165,14 +195,21 @@ export async function expectJsonEqualOnAllPages(
 
 export async function waitUntilEqualOnAllPages(
   pages: [Page, Page],
-  selector: IDSelector
+  selector: IDSelector,
+  options?: {
+    maxTries?: number;
+    interval?: number;
+  }
 ) {
-  for (let i = 0; i < 20; i++) {
+  const maxTries = options?.maxTries ?? 20;
+  const interval = options?.interval ?? 100;
+
+  for (let i = 0; i < maxTries; i++) {
     const [value1, value2] = await getBoth(pages, selector);
     if (_.isEqual(value1, value2)) {
       return; // Great, we're done!
     } else {
-      await sleep(100);
+      await sleep(interval);
     }
   }