+++ examples/nextjs-comments-overlay/src/components/icons/StarIcon.tsx
@@ -1,0 +1,19 @@
+import { ComponentProps } from "react";
+
+export function StarIcon(props: ComponentProps<"svg">) {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        fill="#212134"
+        d="m12 18.26-7.053 3.948 1.575-7.928L.587 8.792l8.027-.952L12 .5l3.386 7.34 8.027.952-5.935 5.488 1.575 7.928L12 18.26Z"
+      ></path>
+    </svg>
+  );
+}