+++ packages/liveblocks-core/src/crdts/LiveRegister.ts
@@ -2,7 +2,8 @@ import type { LiveNode } from "../crdts/Lson";
 import { nn } from "../lib/assert";
 import type { Json } from "../lib/Json";
 import { nanoid } from "../lib/nanoid";
-import type { CreateChildOp, CreateRegisterOp, Op } from "../protocol/Op";
+import { deepClone } from "../lib/utils";
+import type { CreateOp, CreateRegisterOp, Op } from "../protocol/Op";
 import { OpCode } from "../protocol/Op";
 import type { IdTuple, SerializedRegister } from "../protocol/SerializedCrdt";
 import { CrdtType } from "../protocol/SerializedCrdt";
@@ -78,7 +79,7 @@ export class LiveRegister<TValue extends Json> extends AbstractCrdt {
   }
 
   /** @internal */
-  _attachChild(_op: CreateChildOp): ApplyResult {
+  _attachChild(_op: CreateOp): ApplyResult {
     throw new Error("Method not implemented.");
   }
 
@@ -105,5 +106,9 @@ export class LiveRegister<TValue extends Json> extends AbstractCrdt {
   /** @internal */
   _toImmutable(): Immutable {
     return this._data;
+  }
+
+  clone(): TValue {
+    return deepClone(this.data);
   }
 }