+++ packages/liveblocks-react-comments/rollup.config.ts
@@ -5,7 +5,7 @@ import esbuild from "rollup-plugin-esbuild";
 import preserveDirectives from "rollup-plugin-preserve-directives";
 
 import { clean } from "./plugins/rollup/clean";
-import { scss } from "./plugins/rollup/scss";
+import { styles } from "./plugins/rollup/styles";
 
 const SRC_DIR = "src";
 const DIST_DIR = "dist";
@@ -54,18 +54,18 @@ const configs: RollupOptions[] = [
       preserveDirectives(),
 
       // Build .css files
-      scss({
+      styles({
         files: [
           {
-            entry: `${SRC_DIR}/styles/index.scss`,
+            entry: `${SRC_DIR}/styles/index.css`,
             destination: "styles.css",
           },
           {
-            entry: `${SRC_DIR}/styles/dark/media-query.scss`,
+            entry: `${SRC_DIR}/styles/dark/media-query.css`,
             destination: "styles/dark/media-query.css",
           },
           {
-            entry: `${SRC_DIR}/styles/dark/attributes.scss`,
+            entry: `${SRC_DIR}/styles/dark/attributes.css`,
             destination: "styles/dark/attributes.css",
           },
         ],