+++ packages/liveblocks-react-comments/src/components/Thread.tsx
@@ -195,6 +195,7 @@ export const Thread = forwardRef(
                 <Comment
                   key={comment.id}
                   className="lb-thread-comment"
+                  root={false}
                   comment={comment}
                   indentContent={indentCommentContent}
                   showDeleted={showDeletedComments}
@@ -247,6 +248,7 @@ export const Thread = forwardRef(
           {showComposer && (
             <Composer
               className="lb-thread-composer"
+              root={false}
               threadId={thread.id}
               defaultCollapsed={showComposer === "collapsed" ? true : undefined}
               overrides={{